+++ src/Provider/EveOnline.php
@@ -4,15 +4,41 @@ namespace Killmails\OAuth2\Client\Provider;
 
 use League\OAuth2\Client\Provider\AbstractProvider;
 use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
-use League\OAuth2\Client\Provider\ResourceOwnerInterface;
 use League\OAuth2\Client\Token\AccessToken;
 use League\OAuth2\Client\Tool\BearerAuthorizationTrait;
 use Psr\Http\Message\ResponseInterface;
 
 class EveOnline extends AbstractProvider
 {
-
     use BearerAuthorizationTrait;
+
+    /**
+     * @var string Key used in a token response to identify the resource owner.
+     */
+    const ACCESS_TOKEN_RESOURCE_OWNER_ID = 'CharacterID';
+
+    /**
+     * @var string URL path for autorization.
+     */
+    const PATH_AUTHORIZE = '/oauth/authorize';
+
+
+    /**
+     * @var string URL path for token.
+     */
+    const PATH_TOKEN = '/oauth/token';
+
+
+    /**
+     * @var string URL path for user details.
+     */
+    const PATH_USER = '/oauth/verify';
+
+
+    /**
+     * @var string Scope separator
+     */
+    const SCOPE_SEPARATOR = ' ';
 
     /**
      * Domain
@@ -28,7 +54,7 @@ class EveOnline extends AbstractProvider
      */
     public function getBaseAuthorizationUrl()
     {
-        return $this->domain . '/oauth/authorize';
+        return $this->domain . self::PATH_AUTHORIZE;
     }
 
     /**
@@ -40,7 +66,7 @@ class EveOnline extends AbstractProvider
      */
     public function getBaseAccessTokenUrl(array $params)
     {
-        return $this->domain . '/oauth/token';
+        return $this->domain . self::PATH_TOKEN;
     }
 
     /**
@@ -52,7 +78,7 @@ class EveOnline extends AbstractProvider
      */
     public function getResourceOwnerDetailsUrl(AccessToken $token)
     {
-        return $this->domain . '/oauth/verify';
+        return $this->domain . self::PATH_USER;
     }
 
     /**
@@ -67,7 +93,7 @@ class EveOnline extends AbstractProvider
     {
         return [];
     }
-    
+
     /**
      * Returns the string that should be used to separate scopes when building
      * the URL for requesting an access token.
@@ -76,7 +102,7 @@ class EveOnline extends AbstractProvider
      */
     protected function getScopeSeparator()
     {
-        return ' ';
+        return self::SCOPE_SEPARATOR;
     }
 
     /**