+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Changelog
 All Notable changes to `oauth2-eve` will be documented in this file
 
+## 1.0.0 - 2018-05-20
+
+### Added
+- PHP 7.1
+- PHP 7.2
+- Use OAuth 2.0 Client 2.x
+
+### Deprecated
+- PHP 5.5
+
+### Fixed
+- Nothing
+
+## 0.2.2 - 2018-05-19
+
+### Added
+- Nothing
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Use space as `scope` separator instead of comma
+
 ## 0.2.0 - 2016-02-04
 
 ### Added